requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
nltk